Move common logging functions into utils file
